+++ src/parms.c
@@ -15,7 +15,7 @@
 #ifndef MALLOC
 #define MALLOC malloc
 #else
-extern void* MALLOC(int);
+extern void* MALLOC(size_t);
 #endif
 
 STATIC void parmFileError(char *parfile, int line);
